Reverts the wasm work and merges maze back into the server
