--- Cargo.toml
@@ -1,6 +1,0 @@
-[workspace]
-
-members = [
-    "server",
-    "maze",
-]
