--- maze/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "sprokit_maze"
-version = "0.1.0"
-authors = ["Michael Daffin <michael@daffin.io>"]
-edition = "2018"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-derive_more = "0.15.0"
-rand = "0.7.2"
-serde = { version = "1.0", features = ["derive"] }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen = "0.2"
-console_error_panic_hook = "0.1.1"
-wee_alloc = "0.4.2"
-
-[dev-dependencies]
-serde_json = "1.0.41"
-lazy_static = "1.4.0"
