+++ server/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-maze = { path = "../maze", package = "sprokit_maze" }
 actix-http = "0.2.10"
 actix-service = "0.4.2"
 actix-web = "1.0.8"
